[cert-manager] put cloudflare in cert manager ns
It needed to be in this namespace to renew, but I'm not sure
how the cert was ever generated in the first pace if the secret
needed to be there. Maybe it used to be and was refactored,
but since we reload a saved version, this might have been hidden.